fix: Move configs under typed_validator
