--- src/ai/backend/common/configs/BUILD
@@ -1,1 +1,0 @@
-python_sources(name="src")
