+++ src/ai/backend/manager/config/shared.py
@@ -184,12 +184,11 @@ import yarl
 from pydantic import BaseModel, Field, IPvAnyNetwork
 
 from ai.backend.common import config
-from ai.backend.common.configs.types import HostPortPair as HostPortPairModel
-from ai.backend.common.configs.types import TimeZone
 from ai.backend.common.defs import DEFAULT_FILE_IO_TIMEOUT
 from ai.backend.common.etcd import AsyncEtcd, ConfigScopes
 from ai.backend.common.identity import get_instance_id
-from ai.backend.common.typed_validators import TimeDuration
+from ai.backend.common.typed_validators import HostPortPair as HostPortPairModel
+from ai.backend.common.typed_validators import TimeDuration, TimeZone
 from ai.backend.common.types import (
     HostPortPair,
     SlotName,
