+++ tests/manager/conftest.py
@@ -40,9 +40,9 @@ from sqlalchemy.ext.asyncio.engine import AsyncEngine as SAEngine
 
 from ai.backend.common.auth import PublicKey, SecretKey
 from ai.backend.common.config import ConfigurationError
-from ai.backend.common.configs.types import HostPortPair as HostPortPairModel
 from ai.backend.common.lock import FileLock
 from ai.backend.common.plugin.hook import HookPluginContext
+from ai.backend.common.typed_validators import HostPortPair as HostPortPairModel
 from ai.backend.logging import LocalLogger, LogLevel
 from ai.backend.manager.api.context import RootContext
 from ai.backend.manager.api.types import CleanupContext
